Bot-caverne-du-panda/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.6.4",
    "flask>=3.1.2",
]
